requirements.txt
asgiref==3.2.10
boto3==1.15.17
botocore==1.18.17
click==7.1.2
Django==3.1.2
django-storages==1.10.1
djangorestframework==3.12.1
dlib==19.21.0
face-recognition==1.3.0
face-recognition-models==0.3.0
jmespath==0.10.0
numpy==1.19.2
Pillow==8.0.0
pkg-resources==0.0.0
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.10
